requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
selenium>=4.18.0
webdriver-manager>=4.0.1
pdfplumber>=0.10.0
python-docx>=1.1.0
pandas>=2.0.0
langdetect>=1.0.9
